refactor(core): keep validatePropertyNamesOnCreate out of the public export surface (internal propertyNameGuards module)
