+++ src/core/Entity.ts
@@ -1,39 +1,12 @@
 import { IInstance, SerializedData, generateUUID } from './interfaces.js';
 import { stringifyInstance, decodeFunctionValues } from './utils.js';
 import { PropertyInstance } from './Property.js';
+import { validatePropertyNamesOnCreate } from './propertyNameGuards.js';
 import type { ComputationInstance } from './types.js';
 import type { RelationInstance } from './Relation.js';
 import type { ConstraintInstance } from './Constraint.js';
 
 const validNameFormatExp = /^[a-zA-Z0-9_]+$/;
-
-// 与 storage 的 ID_ATTR/ROW_ID_ATTR 对应（core 不能反向依赖 runtime/storage，故用字面量）。
-// relation 额外保留 'source'/'target'（端点虚拟链接属性）。
-export function validatePropertyNamesOnCreate(
-  ownerName: string | undefined,
-  properties: PropertyInstance[] | undefined,
-  kind: 'Entity' | 'Relation'
-) {
-  const reserved = kind === 'Relation' ? ['id', '_rowId', 'source', 'target'] : ['id', '_rowId'];
-  const seen = new Set<string>();
-  for (const property of (properties || [])) {
-    if (reserved.includes(property.name)) {
-      throw new Error(
-        `Property name "${property.name}" on ${kind.toLowerCase()} "${ownerName}" is reserved. ` +
-        (property.name === 'id' || property.name === '_rowId'
-          ? `The framework manages the "${property.name}" column and would silently overwrite your declaration (including any defaultValue/computation on it). Use a different name, e.g. "externalId".`
-          : `"source"/"target" are the relation's endpoint attributes. Use a different name for the relation property.`)
-      );
-    }
-    if (seen.has(property.name)) {
-      throw new Error(
-        `Duplicate property name "${property.name}" on ${kind.toLowerCase()} "${ownerName}". ` +
-        `Property names must be unique per record — duplicates silently keep only the last declaration while all their computations stay registered against one column.`
-      );
-    }
-    seen.add(property.name);
-  }
-}
 
 export interface EntityInstance extends IInstance {
   name: string;
@@ -178,10 +151,7 @@ export class Entity implements EntityInstance {
     if (args.baseEntity && args.inputEntities) {
       throw new Error(`Entity "${args.name}" declares both baseEntity (filtered entity) and inputEntities (merged entity). These are mutually exclusive declaration modes — to filter a merged entity, declare the merged entity first and create a separate filtered entity on top of it.`);
     }
-    // 保留名与重复名：'id'/'_rowId'（对应 storage 的 ID_ATTR/ROW_ID_ATTR）由框架管理，
-    // 用户声明的同名属性会在 schema 编译时被静默覆盖（含其 defaultValue/computation）；
-    // 重复属性名会静默保留最后一个而计算句柄全部注册。二者都是零告警损坏，声明期直接拒绝。
-    // storage 的 DBSetup.validatePropertyNames 是同一守卫的兜底（覆盖 create 之后 push 的属性）。
+    // 保留名（id/_rowId）与重复属性名守卫：见 propertyNameGuards.ts。
     validatePropertyNamesOnCreate(args.name, args.properties, 'Entity');
 
     const instance = new Entity(args, options);
